requirements.txt
streamlit>=1.28.0
spotipy>=2.22.1
requests>=2.31.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
python-dotenv>=1.0.0
sqlite3